fix: use dedicated template directory
