+++ src/components/MDX/Sandpack/templates/vite-react/package-json.json
@@ -1,0 +1,20 @@
+{
+  "name": "vite-react",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "start": "vite dev"
+  },
+  "dependencies": {
+    "html-format": "^1.1.7",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4",
+    "react-error-boundary": "^4.1.1",
+    "rsc-html-stream": "^0.0.7"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react": "^5.1.4",
+    "@vitejs/plugin-rsc": "^0.5.21",
+    "vite": "^7.3.1"
+  }
+}
