+++ src/components/MDX/Sandpack/templates/vite-rsc/package-json.json
@@ -1,0 +1,18 @@
+{
+  "name": "react.dev",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "start": "vite dev"
+  },
+  "dependencies": {
+    "react": "19.2.4",
+    "react-dom": "19.2.4"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.0.0",
+    "@vitejs/plugin-rsc": "^0.5.21",
+    "rsc-html-stream": "^0.0.7",
+    "vite": "^6.0.0"
+  }
+}
