+++ src/components/MDX/Sandpack/templates/vite-rsc/src/framework/entry.browser.jsx
