+++ src/components/MDX/Sandpack/templates/vite-rsc/src/framework/error-boundary.jsx
