+++ src/components/MDX/Sandpack/templates/vite-rsc/src/framework/request.jsx
